Removed TenantDbContext base class.
